Backend: -On directon delete endpoitn a new validatation is added to check this, director does not have movies related, in case this wants to be remove, an action of force delete shoul be added
Frontend:
-Filter od genger on movies table is fixed.
-Promises to post, put,and delete are add to movies's service
-Validattion of fileds requieres is add on bouth forms
-Actions of save and remove register of a movie on fort componete.ts are add

+++ backend/Controllers/DirectorController.cs
@@ -98,8 +98,12 @@ namespace Cruds_Practice.Controllers
             using var trasaction = _db.Database.BeginTransaction();
             try 
             {
+
                 CT_Director? existingDirector = _db.CT_Director.AsNoTracking().FirstOrDefault(d => d.PKDirector == id);
                 if (existingDirector == null) return NotFound(new { message = "Director not found" });
+
+                int moviesRelated = _db.CT_Movie.AsNoTracking().Where(q => q.FKDirector == id).Count();
+                if (moviesRelated > 0) return BadRequest(new { message = $"{existingDirector.Name} has {moviesRelated} movies releated. Delete this movies first. Or contact an Admin." });
 
                 _db.CT_Director.Remove(existingDirector);
                 _db.SaveChanges();
